Add display route icons to stop rows
Added OBAImageViewConfiguration
Moved OBALabelConfiguration into own file
